Do not bind to 0.0.0.0
This can sometimes bind to IPv6, but our Wifi group is IPv4
This leads to the server not accepting any connections, which is bad.

Instead, we pass the proxy the current ConnectionInfo flow, and when the
flow resolves to "Connected" we start the Proxy server.

If the connection info at any point becomes bad, we stop the proxy.
Upon the connection info changing, we restart the proxy server to bind
to the new hostname