Avoid a foreground service
Using START_STICKY, and the fact that WifiP2P persists outside
of our application lifecycle we can do the following

Upon start button click, we call the ServiceLauncher.startForeground
this launches a service after claiming a boolean lock, but
also launches the Wifi Direct and Proxy and notifications and stuff.

The launched Android Service is used with START_STICKY so that it will
restart after Android kills it

The only job the Android Service has is to call ServiceLauncher.startForeground
again

Since the boolean lock is claimed while the process is alive, the ServiceLauncher will
not double-start, it will simply ignore the repeat command.

When restoring the process from kill though, the flag will be unset and will thus launch
the service launcher

Because Wifip2p is outside of our application lifecycle, we re-use the existing network
and continue where we left off before being killed

Additional OS suggestions like claiming a wakelock and ignoring power optimization
can help ensuring the service always re-launches properly.

The service has no destroy() cleanup, since it handles wifip2p which is outside
of the service lifecycle, its improper for us to call cleanup anyway in the service

Also normally wifip2p kills itself, but we can avoid this by simply
not tearing it down when the kill hook fires