Use the URL parser built-in
If our provided line is not a fully valid URL, we force it.
Generally speaking, all lines have at least the host provided, but not
always the protocol.

If the protocol is absent, we assume it's HTTP because what else could
it be. From there, we can parse the URL, and then we can guess or assign
a port and parse the file path